% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{bootpval Function}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{data}

\item{conf.level}{confidence level}

\item{iter}{number of iterations for bootstrapping}

\item{mu0}{null value}

\item{test}{upper tail, lower tail, or two--both--tails}
}
\value{
plot, and list of tcalc, pvalue, sample size, data points, type of test, and conf. interval
}
\description{
bootpval Function
}
\examples{

bootpval(c(4, 5, 2, 3, 4, 5, 1), conf.level = 0.95, mu0 = 1)
}
