% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire damage}
\format{
##
A data frame with 15 rows and 2 columns:
\describe{
  \item{distance}{distance from fire station in miles}
  \item{damage}{fire damage in thousands of dollars}
}
}
\source{
Statistics for Engineering and the Sciences, Sixth Edition by
William Mendenhall
}
\usage{
fire
}
\description{
Residential fire damage data from a fire insurance company
Report ...
}
\keyword{datasets}
