% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{"My" Binomial Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations of simulation}

\item{n}{number of trials}

\item{p}{probability of success}
}
\value{
barplot of relative frequencies for each number as well as a table
}
\description{
"My" Binomial Function
}
\examples{
mybin(iter = 10000, n = 15, p = 0.6)
}
