% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function}
\usage{
myci(x)
}
\arguments{
\item{x}{data}
}
\value{
confidence interval from t test
}
\description{
Confidence Interval Function
}
\examples{
myci(c(1, 5, 10, 2, 4, 6))
}
