% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymlbeta.R
\name{mymlbeta}
\alias{mymlbeta}
\title{Maximum Likelihood Beta Function}
\usage{
mymlbeta(x, alpha, beta, ...)
}
\arguments{
\item{x}{data}

\item{alpha}{shape1 range outlines}

\item{beta}{shape2 outlines}

\item{...}{extra arguments}
}
\value{
List of data, coordinates of maximum likelihood, maximum likelihoods for alpha and beta
}
\description{
Maximum Likelihood Beta Function
}
\examples{
mymlbeta(x = rbeta(20, 5, 5), alpha = seq(0.1, 10, length = 100), beta = seq(0.1, 10, length = 100))
}
