% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My n curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{distribution mean}

\item{sigma}{distribution standard deviation}

\item{a}{x value used to calculate P(Y < a)}
}
\value{
graph of curve with region and area/probability plotted as well as a list of mu, sigma, and a
}
\description{
My n curve
}
\examples{
myncurve(mu = 100, sigma = 15, a = 95)
}
