% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{n tickets Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats}

\item{gamma}{overbooking probability}

\item{p}{probability of show}
}
\value{
list with nd (number of tickets for discrete), nc (number of tickets for continuous), N, gamma, p
}
\description{
n tickets Function
}
\examples{
ntickets(400, 0.02, 0.95)
}
