% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto Function}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{any vector}

\item{mn}{title of barplot}

\item{...}{allows for more arguments}
}
\value{
pareto barplot
}
\description{
Pareto Function
}
\examples{
x <- c(1, 2, 2, 3, 3, 3, 3, 4, 4, 5, 5, 5, 5, 5, 5)
pareto(x)
}
